refactor PC
