+++ pgmpy/estimators/PC.py
@@ -3,7 +3,6 @@
 from itertools import chain, combinations, permutations
 from typing import (
     Callable,
-    Collection,
     Dict,
     FrozenSet,
     Hashable,
@@ -15,18 +14,15 @@ from typing import (
 
 import networkx as nx
 import pandas as pd
-from joblib import Parallel, delayed
-from tqdm.auto import tqdm
 
-from pgmpy import config
 from pgmpy.base import DAG, PDAG, UndirectedGraph
-from pgmpy.estimators import ExpertKnowledge, StructureEstimator
+from pgmpy.estimators import ExpertKnowledge
+from pgmpy.estimators.BaseConstraintEstimator import BaseConstraintEstimator
 from pgmpy.estimators.CITests import get_callable_ci_test
-from pgmpy.global_vars import logger
 from pgmpy.independencies import Independencies
 
 
-class PC(StructureEstimator):
+class PC(BaseConstraintEstimator):
     """
     Class for constraint-based estimation of DAGs using the PC algorithm
     from a given data set.  Identifies (conditional) dependencies in data
@@ -286,239 +282,6 @@ class PC(StructureEstimator):
             raise ValueError(
                 f"return_type must be one of: dag, pdag, cpdag, or skeleton. Got: {return_type}"
             )
-
-    def build_skeleton(
-        self,
-        variant: str = "stable",
-        ci_test: Union[str, Callable, None] = None,
-        significance_level: float = 0.01,
-        max_cond_vars: int = 5,
-        expert_knowledge: Optional[ExpertKnowledge] = None,
-        enforce_expert_knowledge: bool = False,
-        n_jobs: int = -1,
-        show_progress: bool = True,
-        **kwargs,
-    ) -> Tuple[UndirectedGraph, Dict[Tuple[str, str], Set[str]]]:
-        """
-        Estimates a graph skeleton (UndirectedGraph) from a set of independencies
-        using (the first part of) the PC algorithm. The independencies can either be
-        provided as an instance of the `Independencies`-class or by passing a
-        decision function that decides any conditional independency assertion.
-        Returns a tuple `(skeleton, separating_sets)`.
-
-        If an Independencies-instance is passed, the contained IndependenceAssertions
-        have to admit a faithful BN representation. This is the case if
-        they are obtained as a set of d-separations of some Bayesian network or
-        if the independence assertions are closed under the semi-graphoid axioms.
-        Otherwise, the procedure may fail to identify the correct structure.
-
-        Parameters
-        ----------
-
-        Returns
-        -------
-        skeleton: UndirectedGraph
-            An estimate for the undirected graph skeleton of the BN underlying the data.
-
-        separating_sets: dict
-            A dict containing for each pair of not directly connected nodes a
-            separating set ("witnessing set") of variables that makes them
-            conditionally independent. (needed for edge orientation procedures)
-
-        References
-        ----------
-        [1] Neapolitan, Learning Bayesian Networks, Section 10.1.2, Algorithm 10.2 (page 550)
-            http://www.cs.technion.ac.il/~dang/books/Learning%20Bayesian%20Networks(Neapolitan,%20Richard).pdf
-        [2] Koller & Friedman, Probabilistic Graphical Models - Principles and Techniques, 2009
-            Section 3.4.2.1 (page 85), Algorithm 3.3
-        """
-        # Initialize initial values and structures.
-        lim_neighbors = 0
-        separating_sets = dict()
-        ci_test = get_callable_ci_test(
-            ci_test, full=True, data=None
-        )  # this is called twice, before on PC estimate
-
-        if expert_knowledge is None:
-            expert_knowledge = ExpertKnowledge()
-
-        if expert_knowledge.search_space:
-            expert_knowledge.limit_search_space(self.data.columns)
-
-        if show_progress and config.SHOW_PROGRESS:
-            pbar = tqdm(total=max_cond_vars)
-            pbar.set_description("Working for n conditional variables: 0")
-
-        # Step 1: Initialize a fully connected undirected graph
-        graph = nx.complete_graph(n=self.variables, create_using=nx.Graph)
-        temporal_ordering = expert_knowledge.temporal_ordering
-        if enforce_expert_knowledge:
-            graph.remove_edges_from(expert_knowledge.forbidden_edges)
-
-        # Exit condition: 1. If all the nodes in graph has less than `lim_neighbors` neighbors.
-        #             or  2. `lim_neighbors` is greater than `max_conditional_variables`.
-        while not all(
-            [len(list(graph.neighbors(var))) < lim_neighbors for var in self.variables]
-        ):
-            # Step 2: Iterate over the edges and find a conditioning set of
-            # size `lim_neighbors` which makes u and v independent.
-            if variant == "orig":
-                for u, v in graph.edges():
-                    if (enforce_expert_knowledge is False) or (
-                        (u, v) not in expert_knowledge.required_edges
-                    ):
-                        for separating_set in PC._get_potential_sepsets(
-                            u, v, temporal_ordering, graph, lim_neighbors
-                        ):
-                            # If a conditioning set exists remove the edge, store the separating set
-                            # and move on to finding conditioning set for next edge.
-                            if ci_test(
-                                u,
-                                v,
-                                separating_set,
-                                data=self.data,
-                                independencies=self.independencies,
-                                significance_level=significance_level,
-                                **kwargs,
-                            ):
-                                separating_sets[frozenset((u, v))] = separating_set
-                                graph.remove_edge(u, v)
-                                break
-
-            elif variant == "stable":
-                # In case of stable, precompute neighbors as this is the stable algorithm.
-                for u, v in graph.edges():
-                    if (enforce_expert_knowledge is False) or (
-                        (u, v) not in expert_knowledge.required_edges
-                    ):
-                        for separating_set in PC._get_potential_sepsets(
-                            u, v, temporal_ordering, graph, lim_neighbors
-                        ):
-                            # If a conditioning set exists remove the edge, store the
-                            # separating set and move on to finding conditioning set for next edge.
-                            if ci_test(
-                                u,
-                                v,
-                                separating_set,
-                                data=self.data,
-                                independencies=self.independencies,
-                                significance_level=significance_level,
-                                **kwargs,
-                            ):
-                                separating_sets[frozenset((u, v))] = separating_set
-                                graph.remove_edge(u, v)
-                                break
-
-            elif variant == "parallel":
-
-                def _parallel_fun(u, v):
-                    for separating_set in PC._get_potential_sepsets(
-                        u, v, temporal_ordering, graph, lim_neighbors
-                    ):
-                        if ci_test(
-                            u,
-                            v,
-                            separating_set,
-                            data=self.data,
-                            independencies=self.independencies,
-                            significance_level=significance_level,
-                            **kwargs,
-                        ):
-                            return (u, v), separating_set
-
-                results = Parallel(n_jobs=n_jobs)(
-                    delayed(_parallel_fun)(u, v)
-                    for (u, v) in graph.edges()
-                    if (enforce_expert_knowledge is False)
-                    or ((u, v) not in expert_knowledge.required_edges)
-                )
-                for result in results:
-                    if result is not None:
-                        (u, v), sep_set = result
-                        graph.remove_edge(u, v)
-                        separating_sets[frozenset((u, v))] = sep_set
-
-            else:
-                raise ValueError(
-                    f"variant must be one of (orig, stable, parallel). Got: {variant}"
-                )
-
-            # Step 3: After iterating over all the edges, expand the search space by increasing the size
-            #         of conditioning set by 1.
-            if lim_neighbors >= max_cond_vars:
-                logger.info(
-                    "Reached maximum number of allowed conditional variables. Exiting"
-                )
-                break
-            lim_neighbors += 1
-
-            if show_progress and config.SHOW_PROGRESS:
-                pbar.update(1)
-                pbar.set_description(
-                    f"Working for n conditional variables: {lim_neighbors}"
-                )
-
-        if show_progress and config.SHOW_PROGRESS:
-            pbar.update(max_cond_vars - lim_neighbors)
-            pbar.close()
-        return graph, separating_sets
-
-    @staticmethod
-    def _get_potential_sepsets(
-        u: Hashable,
-        v: Hashable,
-        temporal_ordering: Dict[Hashable, int],
-        graph: UndirectedGraph,
-        lim_neighbors: int,
-    ) -> Collection[Tuple]:
-        """
-        Return the temporally consistent superset of separating set of u, v.
-
-        The temporal order (if specified) of the superset can only be smaller
-        ("earlier") than the particular node. The neighbors of 'u' satisfying
-        this condition are returned.
-
-        Parameters
-        ----------
-        u: variable
-            The node whose neighbors are being considered for separating set.
-
-        v: variable
-            The node along with u whose separating set is being calculated.
-
-        temporal_ordering: dict
-            The temporal ordering of variables according to prior knowledge.
-
-        graph: UndirectedGraph
-            The graph where separating sets are being calculated for the edges.
-
-        lim_neighbors: int
-            The maximum number of neighbours (conditioning variables) for u, v.
-
-        Returns
-        --------
-        separating_set: set
-            Set containing the superset of separating set of u, v.
-        """
-        separating_set_u = set(graph.neighbors(u))
-        separating_set_v = set(graph.neighbors(v))
-        separating_set_u.discard(v)
-        separating_set_v.discard(u)
-
-        if temporal_ordering != dict():
-            max_order = min(temporal_ordering[u], temporal_ordering[u])
-            for neigh in list(separating_set_u):
-                if temporal_ordering[neigh] > max_order:
-                    separating_set_u.discard(neigh)
-
-            for neigh in list(separating_set_v):
-                if temporal_ordering[neigh] > max_order:
-                    separating_set_v.discard(neigh)
-
-        return chain(
-            combinations(separating_set_u, lim_neighbors),
-            combinations(separating_set_v, lim_neighbors),
-        )
 
     @staticmethod
     def orient_colliders(
